Completado Readme, y corregidos estilos de Eslint
