+++ .eslintrc.json
@@ -1,0 +1,26 @@
+{
+    "env": {
+        "browser": true,
+        "es6": true
+    },
+    "extends": [
+        "airbnb-base"
+    ],
+    "globals": {
+        "Atomics": "readonly",
+        "SharedArrayBuffer": "readonly"
+    },
+    "parserOptions": {
+        "ecmaVersion": 2018
+    },
+    "rules": {
+        "import/extensions": [
+            "error",
+            {
+              "js": "always",
+              "json": "sometimes",
+              "allowForLoopAfterthoughts": true 
+            }
+          ]
+    }
+}
