+++ README.md
@@ -1,2 +1,10 @@
 # Casa de cambio
+
  Casa de cambio
+
+• HTML
+• CSS
+• Vanilla Javascript
+• PokeAPI (https://pokeapi.co) 
+• Eslint, para analisis, estilos y reglas del codigo.
+
